+++ addons/email_template/doc/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 `trunk (saas-2)`
 ----------------
 
+ - ``mail.compose.message``: added support of ``mail_server_id`` from template
  - Server action update
 
   - added `email` server action type, now entirely based on email templates.